Add FactoriesBaseVisitor
